requirements.txt
Flask
Flask-MySQL-Connector
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
tensorflow==2.13.0
keras==2.13.1
numpy==1.23.5
deepface==0.0.79
opencv-python==4.8.1.78